move memory tools to memories folder
